Move existing implementation of HashSet with chaining to separate folder in preparation for open addressing implementation
