+++ test/dataStructures/hashSet/HashSetTest.java
@@ -1,8 +1,7 @@
 package test.dataStructures.hashSet;
 
 import org.junit.Test;
-import src.algorithms.countingSort.CountingSort;
-import src.dataStructures.hashSet.HashSet;
+import src.dataStructures.hashSet.chaining.HashSet;
 
 import java.util.*;
 import java.util.stream.Collectors;
